+++ src/js/errorHandler.js
@@ -1,0 +1,7 @@
+const resetError = (inputElement, errorElement) => {
+  errorElement.textContent = '';
+  errorElement.style.visibility = 'hidden';
+  inputElement.classList.remove('error');
+};
+
+export { resetError };
